Regression of subsets Yj by subsets Xi

Regression of subsets Yj by subsets Xi for comparing all the explanatory-explained pairs (Xi,Yj). SUCCESSIVE SOLUTIONS

concorgmreg(x, px, y, py, r)

Arguments

x

are the n times p and n times q matrices of p and q centered column

px

A row vector which contains the numbers pi, i=1,...,kx, of the kx subsets xi of x : sum(pi)=sum(px)=p. px is the partition vector of x

y

See x

py

The partition vector of y. A row vector containing the numbers qi for i = 1,...,ky of the ky subsets yi of y : sum(qi)=sum(py)=q.

r

The number of wanted successive solutions

Value

A list with following components:

cx

a n times rmatrix of the r explanatory components

v

is a \(q \times r\) matrix of ky row blocks \(v_i\) (\(q_i \times r\)) of axes in Rqi relative to yi; \(v_i'*v_i = \mbox{Id}\)

varexp

is a kx x ky x r array; for a fixed solution k, the matrix varexp[,,k] contains kxky explained variances obtained by a simultaneous regression of all the yj by all the xi, so the values \(\mbox{rho2}(cx[n*(i-1)+1:n*i,k],y_j*v_j[,k]) var(y_j*v_j[,k])\)

Details

For the first solution, \(\sum_i \sum_j \mbox{rho2}(cx_i[,1],y_j*v_j[,1]) \mbox{var}(y_j*v_j[,1])\) is the optimized criterion. The second solution is calculated from the same criterion, but with \(y_j-y_j*v_j[,1]*v_j[,1]'\) instead of the matrices yj and with orthogonalities for having two by two zero correlated the explanatory components defined for each matrix xi. And so on for the other solutions. One solution k associates kx explanatory components (in cx[,k]) to ky explained components. When kx =1 (px=p), take concoreg function This function uses the concorgm function

References

Hanafi & Lafosse (2004) Regression of a multi-set by another based on an extension of the SVD. COMPSTAT'2004 Symposium

Author

Lafosse, R.

Examples

x<-matrix(runif(50),10,5);y<-matrix(runif(90),10,9)
x<-scale(x);y<-scale(y)
cr<-concorgmreg(x,c(2,3),y,c(3,2,4),2)
cr$varexp[1,1,]
#> [1] 0.3414108 0.3137691